refactor(base-app): move table items interface to form settings module
