+++ apps/base-app/src/features/textProfileForm/model/settingsTableItem.ts
